backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
django==4.2.30
django-anymail==14.0
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mysqlclient==2.2.7
pillow==11.3.0
pyjwt==2.12.1
pyyaml==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
